backstage: move status.tsx and related components to it's own folder
