+++ components/o-teaser/src/tsx/teaser.tsx
@@ -5,7 +5,7 @@ import Headshot from './headshot';
 import Image from './image';
 import {Meta} from './meta';
 import RelatedLinks from './related-links';
-import Status from './status';
+import {Status} from './status/status';
 import {Standfirst} from './standfirst';
 import {Title} from './title';
 import {Video} from './video';
@@ -22,12 +22,12 @@ export function Teaser(props: TeaserProps) {
 				{media(props) === 'video' && <Video {...props} />}
 				{props.showTitle && <Title {...props} />}
 				{props.showStandfirst && <Standfirst {...props} />}
-				{props.showStatus ? <Status {...props} /> : null}
-				{props.showCustomSlot ? <CustomSlot {...props} /> : null}
-				{media(props) === 'headshot' ? <Headshot {...props} /> : null}
+				{props.showStatus && <Status {...props} />}
+				{props.showCustomSlot && <CustomSlot {...props} />}
+				{media(props) === 'headshot' && <Headshot {...props} />}
 			</Content>
-			{media(props) === 'image' ? <Image {...props} /> : null}
-			{props.showRelatedLinks ? <RelatedLinks {...props} /> : null}
+			{media(props) === 'image' && <Image {...props} />}
+			{props.showRelatedLinks && <RelatedLinks {...props} />}
 		</Container>
 	);
 }
